Create Youtube_Web_Sites
This is for Adguard in the Custom Rules for blocking Web Sites used to download youtube videos directly ..!! 
To protect the kids from ...they are getting smart but we are getting smarter..!! If you block youtube pages then they can download videos directly without even searching youtube site...that is bad ...very bad...!! So this list is going to be updated as much as I can...It is still small but efficient. Thank you.